Set global chart display options to override the default ChartDisplayOpts.

The options are added to options using options(taxplore.display.optionname = value). The settings are used to set the initial appearance of all all subsequent charts shown with plot_krona.

Usage

set_chart_display_opts(...)

Arguments

...

Named arguments

Examples

set_chart_opts(hue_range = c(40, 140))